refactor Study constructor, improve test
